% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibCurve.R
\name{calibCurve}
\alias{calibCurve}
\title{Calculates regression curve for AAS}
\usage{
calibCurve(curve, order = 1, badpoint = NULL, intercept = TRUE,
  plot = TRUE)
}
\arguments{
\item{curve}{data frame of numeric vectors named 'Conc' and 'Signal' containing the
concentrations and the signals, respectively}

\item{order}{regression curve order, 1 for linear (default), 2 for quadratic.
More options will be aviliable soon}

\item{badpoint}{numeric vector with the points to be ignored in the regresión}

\item{intercept}{logical. if \code{TRUE} (default) the intercept is calculated normally
insthead of being forced to 0}

\item{plot}{Logical. If \code{TRUE} (default) the calibration data is plotted}
}
\value{
Model of the calibration curve
}
\description{
Calculates regression curve for AAS
}
