% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc2frac.R
\name{conc2frac}
\alias{conc2frac}
\title{Convert signals into concentration by using given model}
\usage{
conc2frac(feed, strip, time = NULL)
}
\arguments{
\item{feed}{numeric vector with feed concentrations}

\item{strip}{numeric vector with strip concentrations}

\item{time}{numeric vector with time of aliquots sampling}
}
\value{
Data frame with the transport proccess information
}
\description{
Convert signals into concentration by using given model
}
