% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentrationcycles}
\alias{concentrationcycles}
\title{Lithium concentration results using a membrane}
\format{A list of 5 data frames with 10 rows and 3 variables:
\describe{
  \item{Time}{Time in hours of each aliquot taken during the experiment}
  \item{Phase}{Phase of corresponding aliquot, Feed or Strip}
  \item{Fraction}{Remaining lithium concentration in the feed solution or
  transported lithium concentration to the strip solution}
}}
\source{
C. Paredes, E. Rodriguez de San Miguel, Polymer inclusion membrane for
  the recovery and concentration of lithium from seawater. Master thesis,
  Universidad Nacional Autónoma de México, México City, México, 2020.
}
\usage{
concentrationcycles
}
\description{
A list of 5 datasets, each of one with the transport data of each cycle in
a concentration experiment of lithium using a polymer inclusion membrane.
}
\keyword{datasets}
