% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curvelithium}
\alias{curvelithium}
\title{External standard calibration curve for lithium in water.}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{Conc}{lithium concentration in the standards, in mg/kg}
  \item{Signal}{emission signal of lithium at 670.8 nm, in arbitrary units}
}}
\source{
C. Paredes, E. Rodriguez de San Miguel, Polymer inclusion membrane for
  the recovery and concentration of lithium from seawater. Master thesis,
  Universidad Nacional Autónoma de México, México City, México, 2020.
}
\usage{
curvelithium
}
\description{
A dataset containing the concentrations and emission signals of aqueous
lithium standards measured by Flame Atomic Emission Spectrometry (FAES) at a
Perkin-Elmer 3100 Atomic Absorption Spectrometer.
}
\keyword{datasets}
