% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrMolRat.R
\name{extrMolRat}
\alias{extrMolRat}
\title{Calculates the}
\usage{
extrMolRat(mass = NULL, ratio = NULL, ex1 = NULL, ex2 = NULL,
  mw1 = 246, mw2 = 348, rho1 = NULL, rho2 = NULL)
}
\arguments{
\item{mass}{total mass of extractants when}

\item{ratio}{desired molar ratio of extractant 1/extractant 2 when}

\item{ex1}{measured mass or volume of extractant 1. If parameter has volume units, density must be provided}

\item{ex2}{measured mass or volume of extractant 2. If parameter has volume units, density must be provided}

\item{mw1}{molecular weight of extractant 1. Default is 246 corresponding to LIX-54}

\item{mw2}{molecular weight of extractant 2. Default is 348 corresponding to Cyanex 923}

\item{rho1}{density of extractant 1}

\item{rho2}{density of extractant 2}
}
\value{
Total mass and molar ratio of given ammount of extractants mixed or the amounts of extractants to take
       when desired molar ratio and total mass is provided.
}
\description{
Depending on which parameters are given,
}
