% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPlotSP.R
\name{multiPlotSP}
\alias{multiPlotSP}
\title{Overlays several single-phase transport profiles}
\usage{
multiPlotSP(trans, phase = "strip", trend = NULL, legend = FALSE,
  xlim = NULL, xbreaks = NULL, ylim = NULL, ybreaks = NULL)
}
\arguments{
\item{trans}{list of data frames generated using \code{conc2frac}. This is the only non-optional
parameter}

\item{trend}{list of nls of ...}

\item{xlim}{numeric vector of limits to be considered for X-axis}

\item{xbreaks}{numeric vector of x-axis breaks}

\item{ylim}{numeric vector of limits to be considered for X-axis}

\item{ybreaks}{numeric vector of x-axis breaks}
}
\value{
plot
}
\description{
Two empirical non-linear regression curves may be fitted. The data must be provided in
}
