% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permcoef.R
\name{permcoef}
\alias{permcoef}
\title{Calculates permeability coefficients}
\usage{
permcoef(trans, conc_0, vol, area, units = c("cm^3", "cm^2", "h"))
}
\arguments{
\item{trans}{dataframe}

\item{vol}{volume}

\item{area}{secondary metal considered}

\item{initial}{logical, can a linear profile of secondary metal be assumed?}
}
\value{
data frame
}
\description{
Permeability coefficients across a membrane as derived from integrated
Fick's law can be obtained from transport data according to the equation
\deqn{\ln{\Bigg(\frac{C_{\ce{Li+}}}{C_{\ce{Li+}}^0}\Bigg)}=
-\frac{P~a}{V}t} where \eqn{C_{\ce{Li+}}} and \eqn{C_{\ce{Li+}}^0} are
lithium concentrarion
}
