% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planelithium}
\alias{planelithium}
\title{Bivariated calibration plane for lithium in prescence of sodium.}
\format{A data frame with 40 rows and 3 variables:
\describe{
  \item{Conc}{lithium concentration in the standards, in mg/kg}
  \item{Signal}{absorbance signal of lithium at 670.8 nm, in absorbance
  units}
  \item{Conc.S}{sodium concentration in the standards, in mg/kg}
}}
\source{
Paredes, C. and Rodríguez de San Miguel, E., Selective lithium extraction
  and concentration from diluted alkaline aqueous media by a polymer
  inclusion membrane and application to seawater, Desalination, Volume 487,
  2020, 114500, https://doi.org/10.1016/j.desal.2020.114500.
}
\usage{
planelithium
}
\description{
A dataset containing the concentrations of lithium and sodium combined
standards and absorbance signals measured by Flame Atomic Absorption
Spectrometry (FAAS) at a Perkin-Elmer 3100 Atomic Absorption Spectrometer.
}
\keyword{datasets}
