% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCCurve.R
\name{plotCCurve}
\alias{plotCCurve}
\title{Plot calibration curve and add regression line}
\usage{
plotCCurve(curve, model = NULL, badpoint = NULL)
}
\arguments{
\item{badpoint}{}
}
\description{
Plot calibration curve and add regression line
}
