% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seawaterLiNaK}
\alias{seawaterLiNaK}
\title{Lithium, sodium and potassium transport profiles across a membrane}
\format{A list of 6 data frames (two for each lithium, sodium, and
        potassium) with 14 or 8 rows and 3 variables:
\describe{
  \item{Time}{Time in hours of each aliquot taken during the experiment}
  \item{Phase}{Phase of corresponding aliquot, Feed or Strip}
  \item{Fraction}{Remaining lithium concentration in the feed solution or
  transported lithium concentration to the strip solution}
}}
\source{
Paredes, C. and Rodríguez de San Miguel, E., Selective lithium extraction
  and concentration from diluted alkaline aqueous media by a polymer
  inclusion membrane and application to seawater, Desalination, Volume 487,
  2020, 114500, https://doi.org/10.1016/j.desal.2020.114500.
}
\usage{
seawaterLiNaK
}
\description{
A list of 6 datasets containing by duplicate the transport profiles for
lithium, sodium, and potassium from a synthetic simplified seawater matrix
using a polymer inclusion membrane selective to lithium. Lithium samples
were taken every 45 minutes during 4.5 hours while sodium and potassium
determinations were made in samples taken every 1.5 hours.
}
\keyword{datasets}
