% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepfactor.R
\name{sepfactor}
\alias{sepfactor}
\title{Calculates...}
\usage{
sepfactor(main, main0, secon, secon0, order = 2)
}
\arguments{
\item{main}{data frame generated using \code{conc2frac}. This is the only non-optional
parameter}

\item{main0}{data frame}

\item{secon}{secondary metal considered}

\item{secon0}{logical, can a linear profile of secondary metal be assumed?}

\item{order}{asdf}
}
\value{
data frame
}
\description{
Calculates...
}
