% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmem-package.R
\docType{package}
\name{transmem-package}
\alias{transmem-package}
\title{transmem: General tools to manipulate and visualize membrane processes data.}
\description{
This package provides tools for treating and visualizing the results
from different experiments involving (selective) transport across membranes.
Transport profiles involving up to three species can be easily obtained
as publication-ready plots, several membrane performance parameters can be
calculated and many experimental results can be easily studied through the
package's graphical representations. The package allows the automatization
of many routinary data-treatment tasks allowing the experimenter to focus
on the more important issues of the research.
}
\references{
C. Paredes, E. Rodriguez de San Miguel, Polymer inclusion membrane for
  the recovery and concentration of lithium from seawater. Master thesis,
  Universidad Nacional Autónoma de México, México City, México, 2020.
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Eduardo Rodriguez de San Miguel, \email{erdsmg@unam.mx}
}
