% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPlot.R
\name{transPlot}
\alias{transPlot}
\title{Fits non-linear trend curve to transport profiles}
\usage{
transPlot(trans, trend = NULL, secondary = FALSE, legend = FALSE,
  xlim = NULL, xbreaks = NULL, ylim = NULL, ybreaks = NULL,
  lin.secon = FALSE, sec.trend = "spline", span = 0.75, size = 2.8)
}
\arguments{
\item{trans}{data frame generated using \code{conc2frac}. This is the only non-optional
parameter}

\item{trend}{nls of ...}

\item{secondary}{secondary metal considered}

\item{xlim}{numeric vector of limits to be considered for X-axis}

\item{xbreaks}{numeric vector of x-axis breaks}

\item{ylim}{numeric vector of limits to be considered for X-axis}

\item{ybreaks}{numeric vector of x-axis breaks}

\item{lin.secon}{logical, can a linear profile of secondary metal be assumed?}
}
\value{
plot
}
\description{
Two empirical non-linear regression curves may be fitted. The data must be provided in
}
