% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPlotWR.R
\name{transPlotWR}
\alias{transPlotWR}
\title{Fits non-linear trend curve to transport profiles}
\usage{
transPlotWR(trans, trend = NULL, secondary = NULL, legend = FALSE,
  xlim = NULL, xbreaks = NULL, ylim = NULL, ybreaks = NULL,
  lin.secon = FALSE)
}
\arguments{
\item{trans}{list of data frames generated using \code{conc2frac}. This is the only non-optional
parameter}

\item{trend}{list of nls of ...}

\item{secondary}{list of secondary metal considered}

\item{xlim}{numeric vector of limits to be considered for X-axis}

\item{xbreaks}{numeric vector of x-axis breaks}

\item{ylim}{numeric vector of limits to be considered for X-axis}

\item{ybreaks}{numeric vector of x-axis breaks}

\item{lin.secon}{logical, can a linear profile of secondary metal be assumed?}
}
\description{
Two empirical non-linear regression curves may be fitted. The data must be provided in
}
