% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transTrend.R
\name{transTrend}
\alias{transTrend}
\title{Fits non-linear trend curve to transport profiles}
\usage{
transTrend(trans, model = "paredes", eccen = 2)
}
\arguments{
\item{trans}{data frame generated using \code{conc2frac}}

\item{eccen}{eccentricity factor for model. onli used if \code{model} is set to \code{'paredes'}}
}
\description{
Two empirical non-linear regression curves may be fitted. The data must be provided in
}
