% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibPlane.R
\name{calibPlane}
\alias{calibPlane}
\title{Calculates regression plane}
\usage{
calibPlane(plane, badpoint = NULL)
}
\arguments{
\item{plane}{data frame of numeric vectors named 'Conc', 'Conc.S' and 'Signal' containing the
concentrations of the interest metal, the secondary metal (interferent)
and the signals, respectively}

\item{badpoint}{numeric vector with the points to be ignored in the regresión}

\item{order}{Not implemented yet... regression curve order, 1 for linear (default),
more options will be aviliable soon}
}
\value{
Model of the calibration plane
}
\description{
Calculates regression plane
}
