% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibPlane.R
\name{calibPlane}
\alias{calibPlane}
\title{Calculates regression plane for external standard calibration.}
\usage{
calibPlane(plane, badpoint = NULL, plot = TRUE, lines = 13,
  theta = -30, phi = 40, xlab = "Species 1", ylab = "Species 2",
  zlab = "Signal", pch = 18, cex = 2)
}
\arguments{
\item{plane}{Data frame of numeric vectors named 'Conc', 'Conc.S' and
'Signal'. The vectors must contain the concentrations of
the interest metal (the one whose concentration in the
samples is to be known), the secondary metal (interferent)
and the signals, respectively.}

\item{badpoint}{Numeric vector with the points to be ignored in the
regresión. This allows the easy elimination of ouliers
whithout losing the stored measurement information.}

\item{plot}{Logical. If \code{TRUE}, the default, the calibration data
is plotted.}

\item{lines}{Number of lines to use in the mesh of the plane in the plot.}

\item{xlab}{Label for Z axis (response).}

\item{ylab}{Label for Z axis (response).}

\item{zlab}{Label for Z axis (response).}

\item{pch}{Plotting symbols available in R .}

\item{cex}{The size of pch symbols.}
}
\value{
Model of the calibration plane
}
\description{
The function calculates the regression plane of external standard
calibration to later use it to convert signals into concentration values.
It differs from \code{\link{calibCurve}} in the number of explanatory
variables which 2 in this case. This function is useful when some
interference effect is being considered (e.g signal magnification of
lithium atomic absortion due to the precesnce of sodium in the sample).
The calibration plane is plotted if desired.
}
\details{
A simple linear method (i.e \code{lm()}) is applied to obtain the
regression equation. Model assumptions (e.g normal distribution of
residuals) must be verified by the user.
}
\examples{
  calibData <- data.frame(Conc = c(0.0, 0.5, 1.0, 2.5, 5.0,
                                   0.0, 0.5, 1.0, 2.5, 5.0,
                                   0.0, 0.5, 1.0, 2.5, 5.0),
                          Conc.S = c(0.0, 0.0, 0.0, 0.0, 0.0,
                                     0.2, 0.2, 0.2, 0.2, 0.2,
                                     0.6, 0.6, 0.6, 0.6, 0.6),
                          Signal = c(0.006, 0.053, 0.120, 0.300, 0.599,
                                     0.018, 0.065, 0.132, 0.312, 0.611,
                                     0.042, 0.089, 0.156, 0.336, 0.635))
  planeModel <- calibPlane(plane = calibData)
  summary(planeModel$model)
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Eduardo Rodriguez de San Miguel, \email{erdsmg@unam.mx}
}
