% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc2frac.R
\name{conc2frac}
\alias{conc2frac}
\title{Creates a data frame to be used as complete selfcontained transport data set}
\usage{
conc2frac(feed, strip, time = NULL, correct.strip = FALSE,
  normalize = TRUE)
}
\arguments{
\item{feed}{Numeric vector with concentrations in the feed phase}

\item{strip}{Numeric vector with concentrations in the strip phase}

\item{time}{Numeric vector with time at which the aliquots were
sampled. It is an optional parameter, if not provided
regular unitary time intervals are assumed.}

\item{correct.strip}{Logical. If \code{FALSE}, the default, the information
about the amount transported to the strip phase is used
as recived but if it is set to \code{TRUE}, the
initial concentration in the strip phase is substracted
to all concentrations in the same phase. This is
particullary useful when the blank signal is
significative or there is background noise.}

\item{normalize}{Logical. If \code{TRUE}, the default, all
concentrations are divided by the initial concentration
in the feed phase to give results in fractions}
}
\value{
Data frame with the transport proccess information
}
\description{
The function transforms data from numeric vectors containing concentrations
in feed and strip solutions to a data frame that contains complete data
of transport process and can be used by several functions inside the
package. Output data frame can contain normalized fractions still remaining
in the feed and already transported to the strip phase, or the original
concentration data provided.
}
\details{
The change in concentration of species in the feed and strip phases as a
function of time are the main magnitudes being measured in processes
involving transport across membranes. The best form to deal with such data
is inside a dataframe containing the information about the amount of given
species in both phases, the phase, and the transport time transcurred.

Usually this function is required after using \code{\link{signal2conc}} to
convert instrumental signals to concentrations.
}
\author{
Cristhian Paredes, \email{craparedesca@unal.edu.co}

Eduardo Rodriguez de San Miguel, \email{erdsmg@unam.mx}
}
