% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentrationcycles}
\alias{concentrationcycles}
\title{Lithium concentration results using a membrane}
\format{A list of 5 data frames with 10 rows and 3 variables:
\describe{
  \item{Time}{Time in hours of each aliquot taken during the experiment}
  \item{Phase}{Phase of corresponding aliquot, Feed or Strip}
  \item{Fraction}{Remaining lithium concentration in the feed solution or
  transported lithium concentration to the strip solution}
}}
\source{
Paredes, C. and Rodríguez de San Miguel, E., Selective lithium extraction
  and concentration from diluted alkaline aqueous media by a polymer
  inclusion membrane and application to seawater, Desalination, Volume 487,
  2020, 114500, https://doi.org/10.1016/j.desal.2020.114500.
}
\usage{
concentrationcycles
}
\description{
A list of 5 datasets, each of one with the transport data of each cycle in
a concentration experiment of lithium using a polymer inclusion membrane.
}
\keyword{datasets}
