% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curvelithium}
\alias{curvelithium}
\title{External standard calibration curve for lithium in water.}
\format{A data frame with 8 rows and 2 variables:
\describe{
  \item{Conc}{lithium concentration in the standards, in mg/kg}
  \item{Signal}{emission signal of lithium at 670.8 nm, in arbitrary units}
}}
\source{
Paredes, C. and Rodríguez de San Miguel, E., Selective lithium extraction
  and concentration from diluted alkaline aqueous media by a polymer
  inclusion membrane and application to seawater, Desalination, Volume 487,
  2020, 114500, https://doi.org/10.1016/j.desal.2020.114500.
}
\usage{
curvelithium
}
\description{
A dataset containing the concentrations and emission signals of aqueous
lithium standards measured by Flame Atomic Emission Spectrometry (FAES) at a
Perkin-Elmer 3100 Atomic Absorption Spectrometer.
}
\keyword{datasets}
