% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSecondary.R
\name{fixSecondary}
\alias{fixSecondary}
\title{Interpolates seccondary species concentration at missed values}
\usage{
fixSecondary(metalConc, time, compTime, order = 1)
}
\arguments{
\item{metalConc}{Seccondary metal concentrations}

\item{time}{Times at which given concentrations were determined}

\item{compTime}{Times at which the seccondary metal must be interpolated}

\item{order}{Order of the polinomia to be fitted to data (1 or 2)}
}
\value{
Vector of interpolated concentrations
}
\description{
If the seccondary species concentration is determinied in just a fraction
of the aliquots and for some reason the concentration in all the aliquots
is required or desired, the function fits a polynomial trend line to the
existing data and interpolates the concentration in missing aliquots.
}
