% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permcoef.R
\name{permcoef}
\alias{permcoef}
\title{calculates...}
\usage{
permcoef(trans, conc_0, vol, area, units = c("cm^3", "cm^2", "h"))
}
\arguments{
\item{area}{secondary metal considered}

\item{initial}{logical, can a linear profile of secondary metal be assumed?}
}
\value{
data frame
}
\description{
calculates...
}
