% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reusecycles}
\alias{reusecycles}
\title{Membrane reuse capability to transport lithium}
\format{A list of 10 data frames with 10 rows and 3 variables:
\describe{
  \item{Time}{Time in hours of each aliquot taken during the cycle}
  \item{Phase}{Phase of corresponding aliquot, Feed or Strip}
  \item{Fraction}{Remaining lithium fraction in the feed solution or
  transported lithium fraction to the strip solution}
}}
\source{
C. Paredes, E. Rodriguez de San Miguel, Polymer inclusion membrane for
  the recovery and concentration of lithium from seawater. Master thesis,
  Universidad Nacional Autónoma de México, México City, México, 2020.
}
\usage{
reusecycles
}
\description{
A list of 10 datasets, each of one with the transport data of each
cycle in a reuse capability experiment of a polimeric inclusion membrane
selective to lithium ions.
}
\keyword{datasets}
