% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal2conc.R
\name{signal2conc}
\alias{signal2conc}
\title{Convert signals into concentration by using given model}
\usage{
signal2conc(signal, model, dilution = NULL, planar = FALSE,
  Conc.S = NULL)
}
\arguments{
\item{signal}{numeric vector of signals to be interpolated}

\item{model}{regression model for the calibration curve}

\item{dilution}{numeric vector of dilution factors applied to samples}

\item{planar}{logical}

\item{Conc.S}{...only used if \code{planar} is set to}
}
\value{
Numeric vector of concentrations
}
\description{
Convert signals into concentration by using given model
}
