% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transTrend.R
\name{transTrend}
\alias{transTrend}
\title{Fits non-linear trend curve to transport profiles}
\usage{
transTrend(trans, model = "paredes", eccen = 2)
}
\arguments{
\item{trans}{data frame generated using \code{conc2frac}}

\item{eccen}{eccentricity factor for model. onli used if \code{model} is set to \code{'paredes'}}
}
\description{
The data must be provided in ... ...
}
\section{Details}{

Two empirical non-linear regression curves may be fitted....
Rodriguez-de-San-Miguel et.al. model, the fractions (\eqn{\Phi}) in feed or
strip phases as a function of time (\eqn{t}) are fitted to \deqn{\Phi=Ae^{-t/d}+y_0}
where \eqn{A}, \eqn{d} and \eqn{y_0} are the parameters to be determined. In this model,
the \eqn{d} parameter determines the steepness of variation, \eqn{y_0} reflects
the limiting value to which the profiles tend to at long pertraction times and \eqn{A}
is not supposed to play an important role in the transport description.
In the Paredes model the equations adjusted to the strip and feed phase are sligthly
different. The transported fraction to the strip phase is
\deqn{\Phi_s=\frac{\alpha_s t^\gamma}{\beta_s^{-1}+t^\gamma}} while the
fraction depleted from the feed phase is
\deqn{\Phi_f=1-\frac{\alpha_f t^\gamma}{\beta_f^{-1}+t^\gamma}}. In those equations \eqn{\alpha}
relates the maximun fraction transported at long pertraction times, \eqn{\beta}
relates the steepness of variation and \eqn{\gamma} is an excentricity factor to improve the adjustment.
The adjustable parameters are \eqn{\alpha} and \eqn{\beta}. The subscripts \eqn{s} and \eqn{f}
relates to strip and feed phases, respectively. If no significant accumulation is presented in the
membrane, the parameters should be quite similar for both phases and a consensus value can be obtained
by various ways. While the aritmetic mean could be appropiate, in order to improve the confidence limit
inherent to the non linear regression fitting, the values can be combinated by using meta-analisys tools.

¿¿EXCENTRICITY FACTOR CAN BE OBTAINED USING NLS??
}

